feat(dishes): save dishes to database and display on screen
- Created CoreData entity 'Dish' to represent dishes
- Implemented logic to save fetched dishes into the CoreData database
- Designed UI to display dishes on screen using SwiftUI
- Fetched dishes from the database and displayed them dynamically in the UI

+++ LittleLemonCapstone/MenuItem.swift
@@ -9,10 +9,11 @@ import Foundation
 
 struct MenuItem: Codable {
     
-    let id: Int64
+    let id: Int
     let title: String
     let description: String
     let price: String
     let image: String
     let category: String
+    
 }
